--- lib/Extension/CodeTransform/Application/Exception/FileAlreadyExists.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Phpactor\Extension\CodeTransform\Application\Exception;
-
-class FileAlreadyExists extends \Exception
-{
-}
