+++ lib/Extension/CodeTransformExtra/Application/ClassInflect.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phpactor\Extension\CodeTransform\Application;
+namespace Phpactor\Extension\CodeTransformExtra\Application;
 
 use Phpactor\CodeTransform\Domain\ClassName;
 use Phpactor\CodeTransform\Domain\SourceCode;
@@ -8,6 +8,7 @@ use Phpactor\Phpactor;
 use Phpactor\Extension\Core\Application\Helper\FilesystemHelper;
 use Phpactor\WorseReflection\Core\Exception\NotFound;
 use Webmozart\Glob\Glob;
+use Phpactor\Extension\CodeTransformExtra\Application\AbstractClassGenerator;
 
 class ClassInflect extends AbstractClassGenerator
 {
