+++ lib/Extension/CodeTransformExtra/Application/Exception/FileAlreadyExists.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Phpactor\Extension\CodeTransformExtra\Application\Exception;
+
+class FileAlreadyExists extends \Exception
+{
+}
