+++ lib/Extension/CodeTransformExtra/CodeTransformExtension.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phpactor\Extension\CodeTransform;
+namespace Phpactor\Extension\CodeTransformExtra;
 
 use Phpactor\CodeBuilder\Adapter\TolerantParser\TolerantUpdater;
 use Phpactor\CodeBuilder\Adapter\Twig\TwigExtension;
@@ -35,24 +35,24 @@ use Phpactor\Extension\SourceCodeFilesystem\SourceCodeFilesystemExtension;
 use Phpactor\Extension\WorseReflection\WorseReflectionExtension;
 use Phpactor\FilePathResolverExtension\FilePathResolverExtension;
 use Phpactor\MapResolver\Resolver;
-use Phpactor\Extension\CodeTransform\Application\ClassInflect;
-use Phpactor\Extension\CodeTransform\Application\ClassNew;
-use Phpactor\Extension\CodeTransform\Application\Transformer;
-use Phpactor\Extension\CodeTransform\Command\ClassInflectCommand;
-use Phpactor\Extension\CodeTransform\Command\ClassNewCommand;
-use Phpactor\Extension\CodeTransform\Command\ClassTransformCommand;
-use Phpactor\Extension\CodeTransform\Rpc\ClassInflectHandler;
-use Phpactor\Extension\CodeTransform\Rpc\ChangeVisiblityHandler;
-use Phpactor\Extension\CodeTransform\Rpc\ClassNewHandler;
-use Phpactor\Extension\CodeTransform\Rpc\ExtractConstantHandler;
-use Phpactor\Extension\CodeTransform\Rpc\ExtractExpressionHandler;
-use Phpactor\Extension\CodeTransform\Rpc\ExtractMethodHandler;
-use Phpactor\Extension\CodeTransform\Rpc\GenerateAccessorHandler;
-use Phpactor\Extension\CodeTransform\Rpc\GenerateMethodHandler;
-use Phpactor\Extension\CodeTransform\Rpc\ImportClassHandler;
-use Phpactor\Extension\CodeTransform\Rpc\OverrideMethodHandler;
-use Phpactor\Extension\CodeTransform\Rpc\RenameVariableHandler;
-use Phpactor\Extension\CodeTransform\Rpc\TransformHandler;
+use Phpactor\Extension\CodeTransformExtra\Application\ClassInflect;
+use Phpactor\Extension\CodeTransformExtra\Application\ClassNew;
+use Phpactor\Extension\CodeTransformExtra\Application\Transformer;
+use Phpactor\Extension\CodeTransformExtra\Command\ClassInflectCommand;
+use Phpactor\Extension\CodeTransformExtra\Command\ClassNewCommand;
+use Phpactor\Extension\CodeTransformExtra\Command\ClassTransformCommand;
+use Phpactor\Extension\CodeTransformExtra\Rpc\ClassInflectHandler;
+use Phpactor\Extension\CodeTransformExtra\Rpc\ChangeVisiblityHandler;
+use Phpactor\Extension\CodeTransformExtra\Rpc\ClassNewHandler;
+use Phpactor\Extension\CodeTransformExtra\Rpc\ExtractConstantHandler;
+use Phpactor\Extension\CodeTransformExtra\Rpc\ExtractExpressionHandler;
+use Phpactor\Extension\CodeTransformExtra\Rpc\ExtractMethodHandler;
+use Phpactor\Extension\CodeTransformExtra\Rpc\GenerateAccessorHandler;
+use Phpactor\Extension\CodeTransformExtra\Rpc\GenerateMethodHandler;
+use Phpactor\Extension\CodeTransformExtra\Rpc\ImportClassHandler;
+use Phpactor\Extension\CodeTransformExtra\Rpc\OverrideMethodHandler;
+use Phpactor\Extension\CodeTransformExtra\Rpc\RenameVariableHandler;
+use Phpactor\Extension\CodeTransformExtra\Rpc\TransformHandler;
 use Twig\Environment;
 use Twig\Loader\ChainLoader;
 use Twig\Loader\FilesystemLoader;
