+++ lib/Extension/CodeTransformExtra/Rpc/GenerateMethodHandler.php
@@ -1,29 +1,28 @@
 <?php
 
-namespace Phpactor\Extension\CodeTransform\Rpc;
+namespace Phpactor\Extension\CodeTransformExtra\Rpc;
 
-use Phpactor\CodeTransform\Domain\Refactor\GenerateAccessor;
+use Phpactor\CodeTransform\Domain\Refactor\GenerateMethod;
 use Phpactor\MapResolver\Resolver;
 use Phpactor\Extension\Rpc\Response\UpdateFileSourceResponse;
 use Phpactor\CodeTransform\Domain\SourceCode;
 use Phpactor\Extension\Rpc\Handler\AbstractHandler;
 
-class GenerateAccessorHandler extends AbstractHandler
+class GenerateMethodHandler extends AbstractHandler
 {
-    const NAME = 'generate_accessor';
+    const NAME = 'generate_method';
     const PARAM_OFFSET = 'offset';
     const PARAM_SOURCE = 'source';
     const PARAM_PATH = 'path';
 
     /**
-     * @var GenerateAccessor
+     * @var GenerateMethod
      */
-    private $generateAccessor;
+    private $generateMethod;
 
-    public function __construct(
-        GenerateAccessor $generateAccessor
-    ) {
-        $this->generateAccessor = $generateAccessor;
+    public function __construct(GenerateMethod $generateMethod)
+    {
+        $this->generateMethod = $generateMethod;
     }
 
     public function name(): string
@@ -42,7 +41,7 @@ class GenerateAccessorHandler extends AbstractHandler
 
     public function handle(array $arguments)
     {
-        $sourceCode = $this->generateAccessor->generateAccessor(
+        $sourceCode = $this->generateMethod->generateMethod(
             SourceCode::fromStringAndPath(
                 $arguments[self::PARAM_SOURCE],
                 $arguments[self::PARAM_PATH]
@@ -50,10 +49,21 @@ class GenerateAccessorHandler extends AbstractHandler
             $arguments[self::PARAM_OFFSET]
         );
 
+        $originalSource = $this->determineOriginalSource($sourceCode, $arguments);
+
         return UpdateFileSourceResponse::fromPathOldAndNewSource(
             $sourceCode->path(),
-            $arguments[self::PARAM_SOURCE],
+            $originalSource,
             (string) $sourceCode
         );
+    }
+
+    private function determineOriginalSource(SourceCode $sourceCode, array $arguments)
+    {
+        $originalSource = $sourceCode->path() === $arguments[self::PARAM_PATH] ?
+            $arguments[self::PARAM_SOURCE] :
+            file_get_contents($sourceCode->path());
+
+        return $originalSource;
     }
 }
