+++ lib/Extension/CodeTransformExtra/Rpc/ImportClassHandler.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phpactor\Extension\CodeTransform\Rpc;
+namespace Phpactor\Extension\CodeTransformExtra\Rpc;
 
 use Phpactor\Extension\Rpc\Response\CollectionResponse;
 use Phpactor\MapResolver\Resolver;
