+++ lib/Extension/CodeTransformExtra/Rpc/RenameVariableHandler.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phpactor\Extension\CodeTransform\Rpc;
+namespace Phpactor\Extension\CodeTransformExtra\Rpc;
 
 use Phpactor\CodeTransform\Domain\Refactor\RenameVariable;
 use Phpactor\MapResolver\Resolver;
