+++ tests/Unit/Extension/CodeTransformExtra/Rpc/ChangeVisiblityHandlerTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Phpactor\Tests\Unit\Extension\CodeTransform\Rpc;
+namespace Phpactor\Tests\Unit\Extension\CodeTransformExtra\Rpc;
 
 use Phpactor\CodeTransform\Domain\Refactor\ChangeVisiblity;
 use Phpactor\CodeTransform\Domain\SourceCode;
-use Phpactor\Extension\CodeTransform\Rpc\ChangeVisiblityHandler;
+use Phpactor\Extension\CodeTransformExtra\Rpc\ChangeVisiblityHandler;
 use Phpactor\Extension\Rpc\Handler;
 use Phpactor\Extension\Rpc\Response\UpdateFileSourceResponse;
 use Phpactor\Tests\Unit\Extension\Rpc\HandlerTestCase;
