+++ tests/Unit/Extension/CoreTransform/Rpc/ExtractExpressionHandlerTest.php
@@ -3,7 +3,7 @@
 namespace Phpactor\Tests\Unit\Extension\CoreTransform\Rpc;
 
 use Phpactor\CodeTransform\Domain\Refactor\ExtractExpression;
-use Phpactor\Extension\CodeTransform\Rpc\ExtractExpressionHandler;
+use Phpactor\Extension\CodeTransformExtra\Rpc\ExtractExpressionHandler;
 use Phpactor\Extension\Rpc\Handler;
 use Phpactor\Extension\Rpc\Response\InputCallbackResponse;
 use Phpactor\CodeTransform\Domain\SourceCode;
