+++ tests/Unit/Extension/CoreTransform/Rpc/GenerateAccessorHandlerTest.php
@@ -6,7 +6,7 @@ use Phpactor\Extension\Rpc\Handler;
 use Phpactor\CodeTransform\Domain\SourceCode;
 use Phpactor\Extension\Rpc\Response\UpdateFileSourceResponse;
 use Phpactor\CodeTransform\Domain\Refactor\GenerateAccessor;
-use Phpactor\Extension\CodeTransform\Rpc\GenerateAccessorHandler;
+use Phpactor\Extension\CodeTransformExtra\Rpc\GenerateAccessorHandler;
 use Phpactor\Tests\Unit\Extension\Rpc\HandlerTestCase;
 
 class GenerateAccessorHandlerTest extends HandlerTestCase
