+++ tests/Unit/Extension/CoreTransform/Rpc/GenerateMethodHandlerTest.php
@@ -4,7 +4,7 @@ namespace Phpactor\Tests\Unit\Extension\CoreTransform\Rpc;
 
 use Phpactor\CodeTransform\Domain\Refactor\GenerateMethod;
 use Phpactor\CodeTransform\Domain\SourceCode;
-use Phpactor\Extension\CodeTransform\Rpc\GenerateMethodHandler;
+use Phpactor\Extension\CodeTransformExtra\Rpc\GenerateMethodHandler;
 use Phpactor\Extension\Rpc\Handler;
 use Phpactor\Extension\Rpc\Response\UpdateFileSourceResponse;
 use Phpactor\Tests\Unit\Extension\Rpc\HandlerTestCase;
