+++ tests/Unit/Extension/CoreTransform/Rpc/OverrideMethodHandlerTest.php
@@ -3,7 +3,7 @@
 namespace Phpactor\Tests\Unit\Extension\CoreTransform\Rpc;
 
 use Phpactor\Tests\Unit\Extension\Rpc\HandlerTestCase;
-use Phpactor\Extension\CodeTransform\Rpc\OverrideMethodHandler;
+use Phpactor\Extension\CodeTransformExtra\Rpc\OverrideMethodHandler;
 use Phpactor\Extension\Rpc\Handler;
 use Phpactor\CodeTransform\Domain\Refactor\OverrideMethod;
 use Phpactor\CodeTransform\Domain\SourceCode as TransformSourceCode;
