+++ tests/Unit/Extension/CoreTransform/Rpc/RenameVariableHandlerTest.php
@@ -2,7 +2,7 @@
 
 namespace Phpactor\Tests\Unit\Extension\CoreTransform\Rpc;
 
-use Phpactor\Extension\CodeTransform\Rpc\RenameVariableHandler;
+use Phpactor\Extension\CodeTransformExtra\Rpc\RenameVariableHandler;
 use Phpactor\CodeTransform\Domain\Refactor\RenameVariable;
 use Phpactor\Extension\Rpc\Handler;
 use Phpactor\Extension\Rpc\Response\InputCallbackResponse;
